homeworks_dl/10_Detection/hw_face_recognition_project/requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
Pillow
facenet-pytorch==2.5.2
tqdm
kaggle